refactor(table): split table rendering responsibilities
