+++ tests/cycles/WF-165/tranche-e-families.ts
@@ -17,6 +17,20 @@ export const TRANCHE_E_FAMILY_MEMBERS = {
   ],
   "packages/bijou/src/core/components/table.ts": [
     "packages/bijou/src/core/components/table.ts",
+    "packages/bijou/src/core/components/table-contract.ts",
+    "packages/bijou/src/core/components/table-model.ts",
+    "packages/bijou/src/core/components/table-measure.ts",
+    "packages/bijou/src/core/components/table-normalize.ts",
+    "packages/bijou/src/core/components/table-fit.ts",
+    "packages/bijou/src/core/components/table-text-ansi.ts",
+    "packages/bijou/src/core/components/table-cell-format.ts",
+    "packages/bijou/src/core/components/table-build.ts",
+    "packages/bijou/src/core/components/table-render-grid.ts",
+    "packages/bijou/src/core/components/table-render-borderless.ts",
+    "packages/bijou/src/core/components/table-render-markdown.ts",
+    "packages/bijou/src/core/components/table-render-expanded.ts",
+    "packages/bijou/src/core/components/table-render-visual.ts",
+    "packages/bijou/src/core/components/table-render-pipe.ts",
   ],
   "packages/bijou-tui/src/app-frame.ts": [
     "packages/bijou-tui/src/app-frame.ts",
